+++ src/refactoring/pages/cart/components/CartList.tsx
@@ -1,6 +1,9 @@
 import { CartItem, Coupon } from '../../../../types';
 import { useDiscountCalculator } from '../../../hooks/useDiscountCalculator';
 import { formatDiscountRate, formatCurrency } from '../../../utils';
+import { CartCard } from './CartCard';
+import { CartCoupon } from './CartCoupon';
+import { CartTotal } from './CartTotal';
 
 interface Props {
   cart: CartItem[];
@@ -39,87 +42,28 @@ export const CartList = ({
     <div>
       <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
       <div className="space-y-2">
-        {cart.map((item) => {
-          const appliedDiscount = getAppliedDiscount(item);
-          return (
-            <div
-              key={item.product.id}
-              className="flex justify-between items-center bg-white p-3 rounded shadow"
-            >
-              <div>
-                <span className="font-semibold">{item.product.name}</span>
-                <br />
-                <span className="text-sm text-gray-600">
-                  {formatCurrency(item.product.price)} x {item.quantity}
-                  {appliedDiscount > 0 && (
-                    <span className="text-green-600 ml-1">
-                      ({formatDiscountRate(appliedDiscount)} 할인 적용)
-                    </span>
-                  )}
-                </span>
-              </div>
-              <div>
-                <button
-                  onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                  className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                >
-                  -
-                </button>
-                <button
-                  onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                  className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                >
-                  +
-                </button>
-                <button
-                  onClick={() => removeFromCart(item.product.id)}
-                  className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                >
-                  삭제
-                </button>
-              </div>
-            </div>
-          );
-        })}
+        {cart.map((item) => (
+          <CartCard
+            key={item.product.id}
+            item={item}
+            getAppliedDiscount={getAppliedDiscount}
+            updateQuantity={updateQuantity}
+            removeFromCart={removeFromCart}
+          />
+        ))}
       </div>
 
-      <div className="mt-6 bg-white p-4 rounded shadow">
-        <h2 className="text-2xl font-semibold mb-2">쿠폰 적용</h2>
-        <select
-          onChange={(e) => applyCoupon(coupons[parseInt(e.target.value)])}
-          className="w-full p-2 border rounded mb-2"
-        >
-          <option value="">쿠폰 선택</option>
-          {coupons.map((coupon, index) => (
-            <option key={coupon.code} value={index}>
-              {coupon.name} -{' '}
-              {coupon.discountType === 'amount'
-                ? `${formatCurrency(coupon.discountValue)}`
-                : `${formatDiscountRate(coupon.discountValue)}`}
-            </option>
-          ))}
-        </select>
-        {selectedCoupon && (
-          <p className="text-green-600">
-            적용된 쿠폰: {selectedCoupon.name}(
-            {selectedCoupon.discountType === 'amount'
-              ? `${formatCurrency(selectedCoupon.discountValue)}`
-              : `${formatDiscountRate(selectedCoupon.discountValue)}`}{' '}
-            할인)
-          </p>
-        )}
-      </div>
+      <CartCoupon
+        coupons={coupons}
+        selectedCoupon={selectedCoupon}
+        applyCoupon={applyCoupon}
+      />
 
-      <div className="mt-6 bg-white p-4 rounded shadow">
-        <h2 className="text-2xl font-semibold mb-2">주문 요약</h2>
-        <div className="space-y-1">
-          <p>상품 금액: {formatCurrency(totalBeforeDiscount)}</p>
-          <p className="text-green-600">할인 금액: {formatCurrency(totalDiscount)}</p>
-          <p className="text-xl font-bold">
-            최종 결제 금액: {formatCurrency(totalAfterDiscount)}
-          </p>
-        </div>
-      </div>
+      <CartTotal
+        totalBeforeDiscount={totalBeforeDiscount}
+        totalDiscount={totalDiscount}
+        totalAfterDiscount={totalAfterDiscount}
+      />
     </div>
   );
 };
