+++ src/refactoring/pages/cart/components/index.ts
@@ -1,3 +1,5 @@
 export * from './ProductList';
 export * from './CartList';
 export * from './ProductCard';
+export * from './CartCard';
+export * from './CartCoupon';
